Add makefile for running nixos-rebuild switch command
This commit adds a new makefile to the project. The makefile contains a single target called "this" which runs the command "nixos-rebuild switch --flake .". This makefile will be used to simplify the process of running the nixos-rebuild command.

Based on the recent user commits and repository commits, it seems that the developer is actively working on improving the project's configuration and documentation.